+++ app/src/main/java/net/canvoki/carburoid/StationListScreen.kt
@@ -3,10 +3,7 @@ package net.canvoki.carburoid
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.widthIn
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
@@ -22,18 +19,16 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import net.canvoki.carburoid.MainSharedViewModel
 import net.canvoki.carburoid.R
-import net.canvoki.carburoid.location.LocationSelector
 import net.canvoki.carburoid.model.GasStation
-import net.canvoki.carburoid.product.CategorizedProductSelector
 import net.canvoki.carburoid.repository.GasStationRepository
 import net.canvoki.carburoid.repository.RepositoryEvent
 import net.canvoki.carburoid.ui.AppScaffold
+import net.canvoki.carburoid.ui.Selectors
 import net.canvoki.carburoid.ui.StationList
 import net.canvoki.carburoid.ui.usermessage.UserMessage
 
@@ -81,9 +76,6 @@ fun StationListScreen(
         }
     }
 
-    val configuration = LocalConfiguration.current
-    val isLandscape = configuration.orientation == android.content.res.Configuration.ORIENTATION_LANDSCAPE
-
     AppScaffold(
         topBar = {
             @OptIn(ExperimentalMaterial3Api::class)
@@ -106,40 +98,14 @@ fun StationListScreen(
             )
         },
     ) {
-        Column(
-            modifier = Modifier.fillMaxSize(),
-        ) {
-            if (isLandscape) {
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    CategorizedProductSelector(
-                        modifier = Modifier.padding(bottom = 8.dp, end = 4.dp).weight(1f),
-                    )
-                    LocationSelector(
-                        modifier = Modifier.padding(bottom = 8.dp).weight(1f),
-                    )
-                }
-            } else {
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    CategorizedProductSelector(
-                        modifier = Modifier.padding(bottom = 8.dp),
-                    )
-                    LocationSelector(
-                        modifier = Modifier.padding(bottom = 8.dp),
-                    )
-                }
-            }
-            StationList(
-                stations = stations,
-                downloading = isDownloading,
-                processing = isProcessing,
-                onRefresh = { repository.launchFetch() },
-                onStationClicked = onStationClicked,
-                modifier = Modifier.weight(1f),
-            )
-        }
+        Selectors()
+        StationList(
+            stations = stations,
+            downloading = isDownloading,
+            processing = isProcessing,
+            onRefresh = { repository.launchFetch() },
+            onStationClicked = onStationClicked,
+            modifier = Modifier.weight(1f),
+        )
     }
 }